[chore] #8: 메뉴 컴포넌트 enum 파일 분리
