+++ app/src/main/java/org/sopt/hyundaicarddive/presentation/ui/detail/component/DetailMenu.kt
@@ -1,7 +1,5 @@
 package org.sopt.hyundaicarddive.presentation.ui.detail.component
 
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -13,39 +11,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import org.sopt.hyundaicarddive.R
+import org.sopt.hyundaicarddive.presentation.type.DetailMenuItem
 import org.sopt.hyundaicarddive.ui.theme.HYUNDAICARDDIVETheme
-
-enum class DetailMenuItem(
-    @DrawableRes val iconRes: Int,
-    @StringRes val stringRes: Int
-) {
-    SPACE(
-        iconRes = R.drawable.ic_spacedetail_space_28,
-        stringRes = R.string.detail_menu_space_title
-    ),
-    BOOKSEARCH(
-        iconRes = R.drawable.ic_spacedetail_booksearch_28,
-        stringRes = R.string.detail_menu_book_search_title
-    ),
-    RESERVATION(
-        iconRes = R.drawable.ic_spacedetail_reservation_28,
-        stringRes = R.string.detail_menu_reservation_title
-    ),
-    REVIEW(
-        iconRes = R.drawable.ic_spacedetail_review_28,
-        stringRes = R.string.detail_menu_review_title
-    ),
-    BOOK(
-        iconRes = R.drawable.ic_spacedetail_book_28,
-        stringRes = R.string.detail_menu_book_title
-    )
-}
 
 @Composable
 fun DetailTabBar() {
